Added recommended menu + dish card component
Added the recommended menu page and the dish card component, so that users will be able to see their recommended menu based on the various meal periods of that day, and also see the specific information of those dishes. I also made rating buttons/components for future need since in out project spec we want the users to be able to rate our menu items and post comments and reviews about them.